Moved MessageHandler from proc to messages.receive.
