+++ src/main/java/com/github/otbproject/otbproject/messages/receive/MessageHandler.java
@@ -1,7 +1,6 @@
-package com.github.otbproject.otbproject.proc;
+package com.github.otbproject.otbproject.messages.receive;
 
 import com.github.otbproject.otbproject.channel.Channel;
-import com.github.otbproject.otbproject.messages.receive.PackagedMessage;
 
 public interface MessageHandler {
     void onMessage(Channel channel, PackagedMessage packagedMessage, boolean timedOut);
